pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm", "ruff"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["betterosi*"]
namespaces = false


[project]
name = "betterosi"
authors = [
    {name = "ika - RWTH Aachen", email = "michael.schuldes@ika.rwth-aachen.de"},
]
description = "betterosi"
readme = "README.md"
requires-python = ">=3.9"
keywords = ["one", "two"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: None",
    "Operating System :: OS Independent",
]


dependencies = [
    'betterproto2[all]',
    'mcap',
    'mcap-protobuf-support',
    'matplotlib',
    'numpy',
    'typer'
]
version = "0.2"

#project.optional-dependencies]
#compile = ['betterproto2_compiler'] #, 'osi3']

[tool.setuptools.package-data]
"betterosi" = [
    "descriptors.json"
]


[project.scripts]
betterosi = "betterosi.viewer:main"
betterosi-tomcap = "betterosi.osi2mcap:app"

[tool.ruff]
lint.ignore = ["E741", "E701"]
